+++ compiler-core/src/language_server/reference.rs
@@ -375,7 +375,7 @@ enum AlternativeVariable {
     Ignore,
 }
 
-struct FindVariableReferences {
+pub struct FindVariableReferences {
     // Due to the structure of some AST nodes (for example, record updates),
     // when we traverse the AST it is possible to accidentally duplicate references.
     // To avoid this, we use a `HashSet` instead of a `Vec` here.
@@ -384,6 +384,27 @@ struct FindVariableReferences {
     definition_location: SrcSpan,
     alternative_variable: AlternativeVariable,
     name: EcoString,
+}
+
+impl FindVariableReferences {
+    pub fn new(variable_definition_location: SrcSpan, variable_name: EcoString) -> Self {
+        Self {
+            references: HashSet::new(),
+            definition_location: variable_definition_location,
+            alternative_variable: AlternativeVariable::Ignore,
+            name: variable_name,
+        }
+    }
+
+    pub fn find_in_module(mut self, module: &TypedModule) -> HashSet<VariableReference> {
+        self.visit_typed_module(module);
+        self.references
+    }
+
+    pub fn find(mut self, expression: &TypedExpr) -> HashSet<VariableReference> {
+        self.visit_typed_expr(expression);
+        self.references
+    }
 }
 
 impl<'ast> Visit<'ast> for FindVariableReferences {